+++ src/main/kotlin/klib/functions/KLibRequire.kt
@@ -1,6 +1,7 @@
 package klib.functions
 
 import klib.kLibInf
+import klib.types.SemVer
 
 /**
  * Check if the used kLib is the required version (Dependency Check)
@@ -14,7 +15,24 @@ import klib.kLibInf
  */
 fun kLibRequire(version: Int) {
     if (kLibInf.versionId != version) {
-        println("This application requires kLib $version! (Version ${kLibInf.versionId} not Compatible)")
+        println("This application requires kLib release $version! (Release ${kLibInf.versionId} not Compatible)")
+        kLibInf.exit(1)
+    }
+}
+
+/**
+ * Check if the used kLib is the required version (Dependency Check)
+ * If not, kLibInf.exit() is called
+ *
+ * @param version
+ * @see kLibInf.exit
+ *
+ * @since 2.1.0
+ * @author Thomas Obernosterer
+ */
+fun kLibRequire(version: SemVer) {
+    if (kLibInf.semver < version) {
+        println("This application requires kLib $version! (Version ${kLibInf.semver} not Compatible)")
         kLibInf.exit(1)
     }
 }
@@ -32,7 +50,7 @@ fun kLibRequire(version: Int) {
  */
 fun kLibRequire(minVersion: Int, maxVersion: Int) {
     if (kLibInf.versionId < minVersion || kLibInf.versionId > maxVersion) {
-        println("This application requires at least kLib $minVersion and at most kLib $maxVersion! (Version ${kLibInf.versionId} not Compatible)")
+        println("This application requires at least kLib release $minVersion and at most kLib release $maxVersion! (Release ${kLibInf.versionId} not Compatible)")
         kLibInf.exit(1)
     }
 }